UI improvements and optimizations, update documentation, improve code readability

+++ IslandCompare/genomeManage/libs/mauvewrapper.py
@@ -1,4 +1,4 @@
-from tempfile import NamedTemporaryFile
+from tempfile import NamedTemporaryFile, mkdtemp
 import subprocess
 import os
 import shutil
@@ -8,11 +8,13 @@ from django.conf import settings
 MAUVE_PATH = settings.MAUVE_PATH
 MAUVE_OUTPUT_PATH = settings.MAUVE_OUTPUT_PATH
 
-def runMauve(sequencepaths, outputbackbonepath):
+def runMauve(sequencepaths, outputbackbonepath, deleteTemp=False):
     # Parameters = path to 2 genbank files
     # Returns None
     # Creates an output file at path outputfile and backbone file at path backbonefile
     scriptFile = NamedTemporaryFile(delete=True)
+    scratchPath1 = mkdtemp()
+    scratchPath2 = mkdtemp()
 
     tmppaths = []
 
@@ -24,8 +26,8 @@ def runMauve(sequencepaths, outputbackbonepath):
 
     with open(scriptFile.name,'w') as script:
         script.write("#!/bin/bash\n")
-        script.write(MAUVE_PATH+" --backbone-output="+outputbackbonepath+
-                     ".backbone ")
+        script.write(MAUVE_PATH+" --output=/dev/null  --scratch-path-1="+scratchPath1+" --scratch-path-1="
+                     + scratchPath2 + " --backbone-output="+outputbackbonepath+".backbone ")
         for sequence in tmppaths:
             script.write(sequence+" ")
         script.close()
@@ -36,8 +38,12 @@ def runMauve(sequencepaths, outputbackbonepath):
     scriptFile.close()
 
     # Delete the temporary gbk files used for mauve
-    for tmp in tmppaths:
-        os.remove(tmp)
+    if deleteTemp:
+        for tmp in tmppaths:
+            os.remove(tmp)
+
+    shutil.rmtree(scratchPath1)
+    shutil.rmtree(scratchPath2)
 
     return None
